Move map-parsing to a separate module: parseMap
